+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed an issue where not all morph targets were being read.
 * Fixed various issues related to loading DLC models and maps for Xenoblade 2 and Xenoblade 3.
 * Fixed an issue where unused alpha channels in glTF diffuse textures would cause black renders in some applications.
+* Fixed an issue where the final field in a uniform buffer struct was not annotated correctly for xc3_shader.
 
 ### Changed
 * Improved rendering accuracy of toon shading pass.